some metadata of product changes
